test: 💍 id は bigint に至るので、マイグレーションを修正した (#37)

+++ db/migrate/20210522033645_create_channels.rb
@@ -1,7 +1,7 @@
 class CreateChannels < ActiveRecord::Migration[6.1]
   def change
     create_table :channels do |t|
-      t.integer :id_number
+      t.bigint :id_number
       t.string :type_name
       t.string :category
       t.string :name
